restructure simlib, add testing
