+++ simlib/src/buffer.cc
@@ -26,7 +26,7 @@
 #include <queue>
 #include <vector>
 
-#include "../actsim_ext.h"
+#include "../../actsim_ext.h"
 
 std::vector<std::unique_ptr<std::queue<expr_res>>> buffers;
