+++ simlib/src/file.cc
@@ -21,7 +21,7 @@
  **************************************************************************
  */
 
-#include "simlib_file.h"
+#include "../simlib_file.h"
 
 #include <common/config.h>
 
@@ -34,7 +34,7 @@
 #include <string>
 #include <unordered_map>
 
-#include "../actsim_ext.h"
+#include "../../actsim_ext.h"
 
 std::unordered_map<size_t, std::pair<std::ifstream, size_t>> input_streams;
 std::unordered_map<size_t, std::pair<std::ofstream, size_t>> output_streams;
