+++ simlib/src/rand_r.cc
