+++ simlib/test/run.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+
+# other kinds of tests here, if required...
+
+./run_actsim_tests.sh
